requirements.txt
langchain==0.3.7
langchain-core
langchain-community
langchain-groq
langchain-pinecone
python-dotenv
streamlit
langchain_google_genai
sentence-transformers
huggingface-hub
openai-whisper
pydub
-e .
